+++ Headers/input_opt_mod.F90
@@ -239,7 +239,7 @@ MODULE Input_Opt_Mod
      LOGICAL                     :: LCONV
      LOGICAL                     :: LTURB
      LOGICAL                     :: LNLPBL
-     LOGICAL                     :: Reconstruct_Conv_Precip_Flux
+     LOGICAL                     :: Grell_Freitas_Convection
      INTEGER                     :: TS_CONV
 
      !----------------------------------------
@@ -760,7 +760,7 @@ CONTAINS
     Input_Opt%LCONV                  = .FALSE.
     Input_Opt%LTURB                  = .FALSE.
     Input_Opt%LNLPBL                 = .FALSE.
-    Input_Opt%Reconstruct_Conv_Precip_Flux = .FALSE.
+    Input_Opt%Grell_Freitas_Convection = .FALSE.
     Input_Opt%TS_CONV                = 0
 
     !----------------------------------------